SiteNavbar: Allow toggling of modals when using the keyboard
Users who prefer to use the keyboard should be allowed to open the
sign up and log in modals.

For this to work only one handler should be present, and not be in
conflict with another one.

By removing the "keyup.enter" handler it allows keyboard users to click
the modal without it being toggled back off right away.

In addition, the logout button does not need a keyup.enter handler
either, as it does the same thing as the click handler.